Make number of line segments between control points configurable in DrawTube effect
Increased default number of line segments by a factor of 3 (to 15), so it should be fine enough.

Parameter is not exposed on the GUI, but it is accessible via editor effect parameters. In the future, if adjustment of this number is often requested by users then we can display it.

Co-authored-by: Andras Lasso <lasso@queensu.ca>