ENH: Allow fast marching to work with merged labelmaps
Before starting the fast marching algorithm, the selected segment is moved to a separate layer.
This ensures that the preview for the segment does not interfere with the other merged segments in the labelmap.

At some point, it may be worth changing the fast marching effect to be an AbstractScriptedSegmentEditorAutoCompleteEffect, so that only the preview would be modified.