ENH: Add alternate threshold effect
New threshold effect is a subclass of the threshold effect in the Slicer core that implements a new flood filling type of threshold.
Once the user has selected a threshold range + optional roi, they can ctrl+click on the image to add the region under the mouse to the image.

The voxels added to the segment are calculated using one of three methods:
* Masking
* GrowCut
* WaterShed