BUG: Fix preview for local threshold
When starting the local threshold effect, the segment preview was not shown.
Fixed by updating the functions in LocalThreshold to call the base class preview() and updatePreviewedSegmentTransparency() to account for the changes to the base threshold effect preview pipeline.
In this effect we now only change the color/opacity of segments after the base class methods have been called.